+++ src/main/java/com/ujc/students/dao/PedidoDaoImpl.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
-public class PedidoDaoImpl extends AbstractDao<Pedido, Long> implements PedidoDao {
+public class PedidoDaoImpl extends AbstractDao<Pedido, Integer> implements PedidoDao {
 
     @Autowired
     private ProdutoDao produtoDao;
@@ -27,7 +27,7 @@ public class PedidoDaoImpl extends AbstractDao<Pedido, Long> implements PedidoDa
     }
 
     @Override
-    public void actualizarEstado(Long pedidoId, EstadoPedido novoEstado) {
+    public void actualizarEstado(Integer pedidoId, EstadoPedido novoEstado) {
         Pedido pedido = findById(pedidoId);
         if (pedido == null)
             throw new IllegalArgumentException("Pedido não encontrado: " + pedidoId);
@@ -36,7 +36,7 @@ public class PedidoDaoImpl extends AbstractDao<Pedido, Long> implements PedidoDa
     }
 
     @Override
-    public void cancelarPedido(Long pedidoId) {
+    public void cancelarPedido(Integer pedidoId) {
         Pedido pedido = findById(pedidoId);
         if (pedido == null)
             throw new IllegalArgumentException("Pedido não encontrado: " + pedidoId);
@@ -47,7 +47,7 @@ public class PedidoDaoImpl extends AbstractDao<Pedido, Long> implements PedidoDa
     }
 
     @Override
-    public List<Pedido> listarPorComprador(Long compradorId) {
+    public List<Pedido> listarPorComprador(Integer compradorId) {
         return createQuery(
             "SELECT p FROM Pedido p WHERE p.comprador.id = ?1", compradorId);
     }
